Extract our Coder interface
